move transpose to matrix_utils
