+++ src/utils/mod.rs
@@ -1,16 +1,1 @@
-use crate::parsing::ast::Parameters;
-
 pub mod matrix_utils;
-
-pub fn transpose(matrix: Vec<Vec<Parameters>>) -> Vec<Vec<Parameters>> {
-    let num_cols = matrix.first().unwrap().len();
-    let mut row_iters: Vec<_> = matrix.into_iter().map(Vec::into_iter).collect();
-    let mut out: Vec<Vec<_>> = (0..num_cols).map(|_| Vec::new()).collect();
-
-    for out_row in out.iter_mut() {
-        for it in row_iters.iter_mut() {
-            out_row.push(it.next().unwrap());
-        }
-    }
-    out
-}
